requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
jinja2==3.1.6
python-multipart==0.0.20
websockets==15.0.1

httpx==0.28.1
google-generativeai==0.8.5

python-telegram-bot==21.11.1

pyautogui==0.9.54
mss==10.0.0
Pillow==11.2.1

playwright==1.51.0

serpapi==0.1.5

pydantic==2.11.1
pydantic-settings==2.8.1
python-dotenv==1.1.0

filelock==3.18.0
